feat(ecs-browser): migrate browser dev components to own function
Move creation of the dev components needed for the Browser to be created outside of the Browser for
easier hooking into
