+++ packages/ecs-browser/src/index.ts
@@ -1,1 +1,2 @@
 export { Browser } from "./Browser";
+export { createBrowserDevComponents } from "./createBrowserDevComponents";
